Name Class
Added a class which uses a provided name parsing strategy to parse the name in the constructor. Defaults to the DyeAndDurham parser if no strategy is provided. Implements the IComparable interface so that names can be sorted as desired. Perhaps this should be done outside of this class so names can be sorted in different ways as desired? Also added a function to concat the first and last names in a standard firstName lastName format for easy output.