Basic Program and Arguments
Added a basic main function which reads in a file from the command line, using the Name class to create a list of names, which are then sorted. Added in a test of this main function which ensures that the function only works when the correct number of arguments are parsed in (1).